+++ cladecombiner/__init__.py
@@ -1,6 +1,6 @@
 from .nomenclature import PangoSc2Nomenclature as PangoSc2Nomenclature
 from .taxon import Taxon as Taxon
+from .taxon_utils import read_taxa as read_taxa
 from .taxonomy_scheme import (
     PhylogeneticTaxonomyScheme as PhylogeneticTaxonomyScheme,
 )
-from .utils import read_taxa as read_taxa
